Increase showInactivity Functionality
I had help from Gabe in order to create a more functional showInactivity command. The command when called tells the person who called it that a list of active members will follow. Then hears if people are sending a message through Discord messages and then responds with a list of the next person who messaged. Then the command stops.